Finished Tron Project
Commented out the code regarding score and food that wasn't needed for the new game. Changed the color and starting position of both snakes.
All of the important changes were made in the Handle Collisions Action file. I created the Leave Tail method that uses raylib's time function to add segments to the snakes.
Then updated the collisions between the segments so that all 4 cases of collision could be detected. And finally changed the game over method so that the segments of both snakes would turn white, and made sure they would stop growing.